+++ src/types/utils/findAll.query.dto.ts
@@ -1,55 +1,16 @@
 import {PaginationDto} from './pagination.dto';
-
-type FindOptionsOrderValue =
-    | 'ASC'
-    | 'DESC'
-    | 'asc'
-    | 'desc'
-    | 1
-    | -1
-    | {
-          direction?: 'asc' | 'desc' | 'ASC' | 'DESC';
-          nulls?: 'first' | 'last' | 'FIRST' | 'LAST';
-      };
-
-type FindOptionsOrderProperty<Property> = Property extends Promise<infer I>
-    ? FindOptionsOrderProperty<NonNullable<I>>
-    : Property extends Array<infer I>
-    ? FindOptionsOrderProperty<NonNullable<I>>
-    : Property extends Function
-    ? never
-    : Property extends Buffer
-    ? FindOptionsOrderValue
-    : Property extends Date
-    ? FindOptionsOrderValue
-    : // : Property extends ObjectID
-    // ? FindOptionsOrderValue
-    Property extends object
-    ? FindOptionsOrder<Property>
-    : FindOptionsOrderValue;
-
-type FindOptionsOrder<Entity> = {
-    [P in keyof Entity]?: P extends 'toString' ? unknown : FindOptionsOrderProperty<NonNullable<Entity[P]>>;
-};
-
-type FindOptionsWherePlainValue<T> = T | 'NULL';
-type FindOptionsWhereOperateObj<T> = {
-    op: 'not';
-    val: FindOptionsWherePlainValue<T>;
-};
-type FindOptionsWhereValue<T> = FindOptionsWherePlainValue<T> | FindOptionsWhereOperateObj<T>;
+import {FindOptionsOrder, FindOptionsWhere, FindOptionsWhereValue} from './find-options';
 
 export class FindAllQueryDto<T> extends PaginationDto {
     relations?: string[];
-    where?: {
-        [P in keyof T]?: T[P] | 'NULL' | {op: 'not'; val: T[P] | 'NULL'};
-    };
+    // where?: {[P in keyof T]?: T[P] | 'NULL' | {op: 'not'; val: T[P] | 'NULL'}};
+    where?: FindOptionsWhere<T>;
     // order?: {[P in keyof T]?: 'ASC' | 'DESC' | 'asc' | 'desc' | 1 | -1};
     order?: FindOptionsOrder<T>;
     updateCounterCacheColumn?: keyof T;
 }
 
-export class FindAllQueryDtoWithKeyword<T> extends FindAllQueryDto<T> {
+export class SearchQueryDto<T> extends FindAllQueryDto<T> {
     keyword?: string;
 }
